requirements.txt
streamlit==1.31.0
pinecone-client==3.0.0
openai
selenium==4.16.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
tiktoken==0.5.2
webdriver-manager==4.0.1